requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
pymysql
pydantic-settings
boto3